tests/wpt/tests/content-security-policy/object-src/object-src-no-url-empty-source-list-blocked.html

<!DOCTYPE html>
<html>

<head>
    <meta http-equiv="Content-Security-Policy" content="object-src; script-src 'self' 'unsafe-inline';">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>

<body>
    <script>
       var t = async_test("Empty object-src source list should block the object and fire a spv");
       window.addEventListener('securitypolicyviolation', t.step_func_done(function(e) {
         assert_equals(e.violatedDirective, "object-src");
       }));
    </script>

    <object type="text/html"></object>
</body>

</html>
